+++ src/Drawables/VanishingSolidRectangle.java
@@ -3,7 +3,7 @@ import java.awt.Color;
 public class VanishingSolidRectangle extends SolidRectangle
 {
 
-    public VanishingSolidRectangle(int x, int y, int width, int height, Color color, GameContent game)
+    public VanishingSolidRectangle(double x, double y, double width, double height, Color color, GameContent game)
     {
         super(x, y, width, height, color, game);
     }
